Implement traceBlock
